refactor: userId를 Long 자료형으로 변경

+++ src/main/java/forif/univ_hanyang/apply/dto/request/AcceptRequest.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 @Data
 public class AcceptRequest {
-    private Set<Integer> applierIds;
+    private Set<Long> applierIds;
     private String applyStatus;
     private Integer studyId;
 }
