+++ src/main/java/forif/univ_hanyang/attendance/StudyAttendanceRepository.kt
@@ -5,6 +5,6 @@ import forif.univ_hanyang.attendance.model.StudyAttendanceId
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface StudyAttendanceRepository : JpaRepository<StudyAttendance, StudyAttendanceId> {
-    fun findByIdUserId(userId: Int): List<StudyAttendance>
+    fun findByIdUserId(userId: Long): List<StudyAttendance>
     fun findByIdStudyId(studyId: Int): List<StudyAttendance>
 }
