+++ src/main/java/forif/univ_hanyang/study/repository/MentorStudyRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MentorStudyRepository extends JpaRepository<MentorStudy, Integer> {
-    List<MentorStudy> findAllById_MentorId(Integer mentorId);
+public interface MentorStudyRepository extends JpaRepository<MentorStudy, Long> {
+    List<MentorStudy> findAllById_MentorId(Long mentorId);
     void deleteAllById_StudyId(Integer studyId);
-    void deleteAllById_MentorId(Integer mentorId);
+    void deleteAllById_MentorId(Long mentorId);
 
 }
